+++ tests/sponsor.spec.ts
@@ -1,7 +1,16 @@
 import {describe, expect, test} from '@jest/globals'
-import {sponsorClient} from './utils'
 import {WhitelistType} from '../src'
-import {POLICY_UUID, ACCOUNT_ADDRESS, CONTRACT_METHOD} from './env'
+import {POLICY_UUID, ACCOUNT_ADDRESS, CONTRACT_METHOD, TOKEN_CONTRACT_ADDRESS, CHAIN_ID, RECIPIENT_ADDRESS, PRIVATE_POLICY_UUID} from './env'
+import {ethers} from 'ethers'
+import {IsSponsorableOptions, SendRawTransactionOptions} from '../src/paymasterclient'
+import {
+  sponsorClient,
+  wallet,
+  tokenAbi,
+  transformIsSponsorableResponse,
+} from './utils'
+
+let TX_HASH = ''
 
 /**
  * Test suite for Sponsor API methods involving whitelist management and spend data retrieval.
@@ -192,4 +201,54 @@ describe('sponsorQuery', () => {
       console.log('Re-addition to FromAccountWhitelist response:', res)
     })
   })
+
+
+    /**
+   * Test for checking if a private policy transaction is sponsorable.
+   */
+    describe('isSponsorable', () => {
+      test('should successfully determine if transaction is sponsorable', async () => {
+        const tokenContract = new ethers.Contract(TOKEN_CONTRACT_ADDRESS, tokenAbi, wallet)
+        const tokenAmount = ethers.parseUnits('0', 18)
+        const nonce = await sponsorClient.getTransactionCount(wallet.address, 'pending')
+  
+        const transaction = await tokenContract.transfer.populateTransaction(RECIPIENT_ADDRESS.toLowerCase(), tokenAmount)
+        transaction.from = wallet.address
+        transaction.nonce = nonce
+        transaction.gasLimit = BigInt(100000)
+        transaction.chainId = BigInt(CHAIN_ID)
+        transaction.gasPrice = BigInt(0)
+  
+        const safeTransaction = {
+          ...transaction,
+          gasLimit: transaction.gasLimit.toString(),
+          chainId: transaction.chainId.toString(),
+          gasPrice: transaction.gasPrice.toString(),
+        }
+  
+        console.log('Prepared transaction:', safeTransaction)
+
+        const opt: IsSponsorableOptions = {
+          PrivatePolicyUUID: PRIVATE_POLICY_UUID
+        };
+
+        const resRaw = await sponsorClient.isSponsorable(safeTransaction, opt)
+        const res = transformIsSponsorableResponse(resRaw)
+        expect(res.Sponsorable).toEqual(true)
+  
+        const txOpt: SendRawTransactionOptions = {
+          PrivatePolicyUUID: PRIVATE_POLICY_UUID,
+          UserAgent: "TEST USER AGENT"
+        };
+
+        const signedTx = await wallet.signTransaction(safeTransaction)
+        try {
+          const tx = await sponsorClient.sendRawTransaction(signedTx,txOpt)
+          TX_HASH = tx
+          console.log('Transaction hash received:', TX_HASH)
+        } catch (error) {
+          console.error('Transaction failed:', error)
+        }
+      }, 100000) // Extends the default timeout as this test involves network calls
+    })
 })
